+++ NEWS.md
@@ -1,5 +1,10 @@
 # Release notes
 
+## 1.0.1 -- not released yet
+
+* Fixed a bug where Granite was missing from the category of non-stone
+  common rocks.
+
 ## 1.0.0 -- 2024-01-28
 
 * Initial release.